+++ react_on_rails_pro/lib/react_on_rails_pro/request.rb
@@ -17,11 +17,9 @@ module ReactOnRailsPro
         perform_request(path, form_with_code(js_code, send_bundle))
       end
 
-      def upload_asset(asset_path, content_type)
-        Rails.logger.info { "[ReactOnRailsPro] Uploading asset #{asset_path}" }
-        raise ReactOnRails::Error, "Asset not found #{asset_path}" unless File.exist?(asset_path)
-
-        perform_request("/upload-asset", form_with_asset(asset_path, content_type))
+      def upload_assets
+        Rails.logger.info { "[ReactOnRailsPro] Uploading assets" }
+        perform_request("/upload-assets", form_with_assets_and_bundle)
       end
 
       def asset_exists_on_vm_renderer?(filename)
@@ -40,8 +38,8 @@ module ReactOnRailsPro
         begin
           response = connection.request(Net::HTTP::Post::Multipart.new(path, form))
         rescue StandardError => e
-          raise ReactOnRailsPro::Error, "Can't connect to VmRenderer renderer.\n"\
-                "Original error:\n#{e}"
+          raise ReactOnRailsPro::Error, "Can't connect to VmRenderer renderer: #{path}.\n"\
+                "Original error:\n#{e}\n#{e.backtrace}"
         end
 
         Rails.logger.info { "[ReactOnRailsPro] VM renderer responded" }
@@ -60,17 +58,44 @@ module ReactOnRailsPro
         form = common_form_data
         form["renderingRequest"] = js_code
         if send_bundle
+          renderer_bundle_file_name = ReactOnRailsPro::ServerRenderingPool::VmRenderingPool.renderer_bundle_file_name
           form["bundle"] = UploadIO.new(
             File.new(ReactOnRails::Utils.server_bundle_js_file_path),
-            ReactOnRails::Utils.server_bundle_js_file_path
+            "application/javascript",
+            renderer_bundle_file_name
           )
+
+          populate_form_with_assets_to_copy(form)
         end
         form
       end
 
-      def form_with_asset(path, content_type)
+      def populate_form_with_assets_to_copy(form)
+        if ReactOnRailsPro.configuration.assets_to_copy.present?
+          ReactOnRailsPro.configuration.assets_to_copy.each_with_index do |asset_path, idx|
+            Rails.logger.info { "[ReactOnRailsPro] Uploading asset #{asset_path}" }
+            raise ReactOnRails::Error, "Asset not found #{asset_path}" unless File.exist?(asset_path)
+
+            content_type = ReactOnRailsPro::Utils.mine_type_from_file_name(asset_path)
+
+            # File.new is very important so that UploadIO does not have confusion over a Pathname
+            # vs. a file path. I.e., Pathname objects don't work.
+            form["assetsToCopy#{idx}"] = UploadIO.new(File.new(asset_path), content_type, asset_path)
+          end
+        end
+        form
+      end
+
+      def form_with_assets_and_bundle
         form = common_form_data
-        form["asset"] = UploadIO.new(path, content_type)
+        populate_form_with_assets_to_copy(form)
+
+        src_bundle_path = ReactOnRails::Utils.server_bundle_js_file_path
+        raise ReactOnRails::Error, "Bundle not found #{src_bundle_path}" unless File.exist?(src_bundle_path)
+
+        renderer_bundle_file_name = ReactOnRailsPro::ServerRenderingPool::VmRenderingPool.renderer_bundle_file_name
+        form["bundle"] = UploadIO.new(File.new(src_bundle_path), "application/javascript",
+                                      renderer_bundle_file_name)
         form
       end